% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/names.R
\name{clean_names}
\alias{clean_names}
\title{Make syntactically valid names and reuturn an object}
\usage{
clean_names(x)
}
\arguments{
\item{x}{an object with names to change}
}
\value{
an object.
}
\description{
Apply \code{\link{make_names}} to object and return an object.
}

