% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/times.R
\name{clean_times}
\alias{clean_times}
\title{Convert Statfin time variable(s) to numeric or date}
\usage{
clean_times(x, time_format = NULL, year_col = "Vuosi",
  sub_year_col = NULL)
}
\arguments{
\item{x}{A data.frame.}

\item{time_format}{"num" for numeric (default for yearly) and
"date" for \code{\link{Date}} (default otherwise)}

\item{year_col}{A column name for year}

\item{sub_year_col}{A column name for quarter or mounth.
If \code{NULL} (default) tries to guess.}
}
\value{
A data.frame
}
\description{
Convert Statfin time variable(s) to numeric or date
}
\examples{
public_debt2 <- clean_times(public_debt)
str(public_debt2)
}

