% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_times2.R
\name{clean_times2}
\alias{clean_times2}
\title{Convert Statfin time variable to Date.}
\usage{
clean_times2(x, time_col = NULL)
}
\arguments{
\item{x}{A data.frame}

\item{time_col}{column in x containing the time variable}
}
\value{
A data.frame
}
\description{
Convert Statfin time variable to Date.
}
\examples{

data_m <- data.frame(values = rnorm(24),
                     Kuukausi = as.vector(sapply(as.character(2016:2017),
                                                 paste0,
                                                 paste0("M", c(paste0("0", 1:9), as.character(10:12))))))
clean_times2(data_m)

data_q <- data.frame(values = rnorm(12),
                     Vuosineljannes = as.vector(sapply(as.character(2016:2018),
                                                       paste0,
                                                       paste0("Q", as.character(1:4)))))
clean_times2(data_q)

data_y <- data.frame(values = rnorm(10),
                     Vuosi = as.character(2010:2019))
clean_times2(data_y)

}
