% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_code}
\alias{extract_code}
\title{Extract only numeric component}
\usage{
extract_code(x)
}
\arguments{
\item{x}{A character vector (or a factor).}
}
\description{
This uses a regular expression to strip all non-numeric character
from a string, including minus sign. Then coerces the result to a number.
Modification of \code{\link[dplyr]{extract_numeric}}, that does not
strip minus sign. Useful to extract numeric codes from code-name variables.
}
\examples{
extract_code("508 Mantta-Vilppula")
}

