% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_name}
\alias{extract_name}
\title{Extract only a non-code component}
\usage{
extract_name(x)
}
\arguments{
\item{x}{A character vector (or a factor).}
}
\value{
A character vector (or a factor).
}
\description{
This uses a regular expression to strip a code from beging of a character
string. Useful to extract names from code-name variables.
}
\examples{
 extract_name("S1311 Valtionhallinto")
}
\seealso{
\code{\link{extract_code}}
}
