% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{loess_series}
\alias{loess_series}
\title{Create trend using local polynomial regression}
\usage{
loess_series(x, time, ...)
}
\arguments{
\item{x}{a numeric vector to smooth.}

\item{time}{a date vector.}

\item{...}{additional arguments such as 'span' that the stats::loess function
takes}
}
\value{
numerical vector
}
\description{
Create trend using local polynomial regression
}
