% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/names.R
\name{make_names}
\alias{make_names}
\title{Make syntactically valid names}
\usage{
make_names(x)
}
\arguments{
\item{x}{a character vector with names to change}
}
\value{
a character vector.
}
\description{
Try to make more convinient names than \code{\link{make.names}}.
}
\details{
Replacements:
 * replace some umlaut mark chracters with a character without (\u00C4 to A)
 * a space with "_"
 * \code{\link{make.names}}
}
\examples{
make_names(c("M\\u00E4\\u00E4r\\u00E4", "Regional code"))
}

