% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_urls.R
\name{open_statfi_qui}
\alias{open_statfi_qui}
\title{Open API url in QUI PxWeb}
\usage{
open_statfi_qui(x = "", db_list_name = NULL)
}
\arguments{
\item{x}{character, url or table code. If table code, set db_list_name}

\item{db_list_name}{character, database of the table}
}
\description{
Can be also used with table codes with \code{\link[pttdatahaku]{pttdatahaku}}.
}
\examples{

  open_statfi_qui("https://statfin.stat.fi/PXWeb/api/v1/fi/StatFin/muutl/statfin_muutl_pxt_119z.px")

}
