% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{pp}
\alias{pp}
\title{Calculate previous years price series}
\usage{
pp(cp, fp, time)
}
\arguments{
\item{cp}{a current price variable.}

\item{fp}{a fixed year price variable.}

\item{time}{a variable for years.}
}
\value{
an vector
}
\description{
For a yearly data calculates a serie in previous year's prices
from fixed and current price series.
}
\examples{
  pp(c(100, 120), c(100, 110), time = c(1,2))

}
