% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{px_code_name}
\alias{px_code_name}
\title{Get code name mapping from pxweb_data}
\usage{
px_code_name(px_data)
}
\arguments{
\item{px_data}{A pxweb_data object.}
}
\value{
A named (column codes) list of named (codes) vectors (names).
}
\description{
Get code name mapping from pxweb_data
}
