% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_na.R
\name{randomize_na}
\alias{randomize_na}
\title{Imputes NAs by randomization}
\usage{
randomize_na(df, from = 1:4)
}
\arguments{
\item{df}{data.frame, the data containing the NAs to be imputed.}

\item{from}{A vector containing the possible imputed values. Defaults to \code{1:4}}
}
\value{
data.frame Returns a data.frame with NAs in numeric columns imputed by randomization.
}
\description{
Imputes NAs by randomization
}
