% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_get_classifications.R
\name{sf_get_class_key}
\alias{sf_get_class_key}
\title{Download data.frame of Statistics Finland classification key.}
\usage{
sf_get_class_key(classification, year)
}
\arguments{
\item{classification}{a regional classification.
Allowed at least: "kuntaryhmitys", "tyossakayntial".}

\item{year}{A year of codes.}
}
\value{
A data.frame.
}
\description{
Download data.frame of classification key from Statistics Finland. At least
regional classifications other that administrative. For administrative
region keys see \code{\link{sf_get_reg_key}}.
}
\details{
Versions for all years are not available and Statistical Finland might also
remove older versions.
}
\examples{
x <- sf_get_class_key("kuntaryhmitys", 2016)
y <- sf_get_class_key("tyossakayntial", 2016)
}
\seealso{
Other sf_class: \code{\link{sf_get_class}},
  \code{\link{sf_get_reg_key}}
}
