% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_ex_munic.R
\name{sf_get_ex_munic}
\alias{sf_get_ex_munic}
\title{Download data.frame for abolished municipalities}
\usage{
sf_get_ex_munic(
 
    url = "http://www.tilastokeskus.fi/meta/luokitukset/_linkki/lakkautetut_kunnat_aakkosissa_15.txt"
)
}
\arguments{
\item{url}{a url for text file in Statistics Finland web page.}
}
\value{
data.frame with columns
   old_code, old_name, "data", new_code, new_name.
}
\description{
Download data.frame for abolished municipalities
}
\examples{
   k <- sf_get_ex_munic()
}
