% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_get_classifications.R
\name{sf_get_reg_key}
\alias{sf_get_reg_key}
\title{Download a data.frame of key to trasform from kunta to other classifications}
\usage{
sf_get_reg_key(to_reg, year)
}
\arguments{
\item{to_reg}{a regional classification to tranforms.
Allowed at least: "maakunta", "skunta".}

\item{year}{A year of codes.}
}
\value{
A data.frame.
}
\description{
Download a data.frame of regional classifications keys from Statistics Finland.
}
\examples{
x <- sf_get_reg_key("maakunta", 2014)
y <- sf_get_reg_key(to_reg = "skunta", year = 2013)
}
\seealso{
Other sf_class: 
\code{\link{sf_get_class_key}()},
\code{\link{sf_get_class}()}
}
\concept{sf_class}
