% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_get_classifications.R
\name{sf_get_reg_keytable}
\alias{sf_get_reg_keytable}
\title{Get regional classification table from Statistics Finland}
\usage{
sf_get_reg_keytable(
  classes = NULL,
 
    url = "http://www.tilastokeskus.fi/static/media/uploads/meta/luokitukset/kooste_2016_kaikki_kielet.xlsx",
  as_factors = default.stringsAsFactors()
)
}
\arguments{
\item{classes}{a name or vector of names of the regional classification.
NULL (default) return all classifications. Names can be names in
orginal table or valid names given by \code{\link{make_names}}.}

\item{url}{a url of the table.}

\item{as_factors}{a logical. Should columns be converted to factors. The
default is from \code{\link{options}}("stringsAsFactors").}
}
\value{
a tibble with columns Knro (classification number of a municipality),
        Kunta (a name of a municipality) and colums for a code and a name of
        the \code{classes} classification.
}
\description{
Get regional classification table for municipalities from Statistics Finland.
Table includes all municipality based regional classifications in Finland.
Classifications include Finnish, Swedish and English versions. The table is
from \url{}
}
\examples{
  key_table <- sf_get_reg_keytable()
  key_table_mk <- sf_get_reg_keytable("Maakunta")
  key_table_mk_sk <- sf_get_reg_keytable(c("Maakunta", "Seutukunta"))
  key_table_mk_sk <- sf_get_reg_keytable(c("Maakunta", "Region"))
}
