% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_ex_munic.R
\name{sf_recode_ex_munic}
\alias{sf_recode_ex_munic}
\title{Replace codes of abolished municipalities with present codes}
\usage{
sf_recode_ex_munic(x, year = NULL, mcodes = sf_get_ex_munic())
}
\arguments{
\item{x}{a vector of municipal codes to replace.}

\item{year}{a year which codes should be used for new codes. Changes up to
1.1. at that year.
`NULL` (the default) uses
the most up to date codes is mcodes.}

\item{mcodes}{a data.frame in form given by \code{\link{sf_get_ex_munic}}.}
}
\description{
Replace codes of abolished municipalities with present codes
}
