% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_test_class.R
\name{sf_test_class}
\alias{sf_test_class}
\title{Test if classification used is Statistics Finland classification}
\usage{
sf_test_class(x, col, ...)
}
\arguments{
\item{x}{avector to test for classification}

\item{col}{a column name from a classification data.frame to test against.
Names form Statistics Finland, see what code{\link{sf_get_class}} gives.}

\item{...}{parameters to select a classification. Passed to \code{\link{sf_get_class}}.}
}
\value{
a locigal. And a warning message with levels that differ.
}
\description{
Test if classification used is Statistics Finland classification
}
