% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_urls.R
\name{statfi_parse_qui_url}
\alias{statfi_parse_qui_url}
\title{Parse Statfi pxweb qui url from api url}
\usage{
statfi_parse_qui_url(url)
}
\arguments{
\item{url}{character, url}
}
\value{

}
\description{
Parse Statfi pxweb qui url from api url
}
\examples{

  statfi_parse_qui_url("StatFin__muutl/statfin_muutl_pxt_119z.px/")

}
