% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_urls.R
\name{statfi_url}
\alias{statfi_url}
\title{Statfi url}
\usage{
statfi_url(
  ...,
  with_base_url = TRUE,
  .base_url = "https://pxnet2.stat.fi/PXWeb/api/v1/fi"
)
}
\arguments{
\item{...}{Character vectors.}

\item{with_base_url}{logical, whether to add the base of url to the output.
Defaults to \code{TRUE}}

\item{.base_url}{A base url for statfi.}
}
\description{
Gives full statfi url
}
\examples{
  statfi_url("StatFin", "kou/vkour/statfin_vkour_pxt_12bq.px")

}
