% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfitools-package.R
\docType{package}
\name{statfitools-package}
\alias{statfitools}
\alias{statfitools-package}
\title{statfitools: Helper Functions for Working with the Data from Statistics Finland}
\description{
Helper functions for working with data from Statistics Finland.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/jhuovari/statfitools}
  \item Report bugs at \url{http://github.com/jhuovari/statfitools/issues}
}

}
\author{
\strong{Maintainer}: Janne Huovari \email{janne.huovari@ptt.fi}

}
\keyword{internal}
