% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_times2.R
\name{time2year}
\alias{time2year}
\title{Convert a time column to year column}
\usage{
time2year(x)
}
\arguments{
\item{x}{A data.frame}
}
\description{
A time named column in a data.frame that is a Date object is converted to year columnt that
is a numeric column.
}
\examples{

x <- data.frame(time = as.Date(c("2020-01-01", "2021-01-01", "2022-01-01")), a = c(1,2,3))
time2year(x)

}
