% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{clean_names}
\alias{clean_names}
\title{Make syntactically valid names and reuturn an object}
\usage{
clean_names(x, to_lower = FALSE)
}
\arguments{
\item{x}{an object with names to change.}

\item{to_lower}{a locigal whether to turn names to lower case.}
}
\value{
an object.
}
\description{
Apply \code{\link{make_names}} with \code{unique = TRUE}
to names of the object and return the object.
}
\examples{
   x <- data.frame("col 1!" = c(1,2), "col 2?" = c(2,4), check.names = FALSE)
   names(clean_names(x))
}
