% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{clean_times}
\alias{clean_times}
\title{Convert Statfin time variable(s) to numeric or date}
\usage{
clean_times(x, time_format = NULL, year_col = "Vuosi",
  sub_year_col = NULL, agg_time = NULL)
}
\arguments{
\item{x}{A data.frame.}

\item{time_format}{"num" for numeric (default for yearly) and
"date" for \code{\link{Date}} (default otherwise)}

\item{year_col}{A column name for year}

\item{sub_year_col}{A column name for quarter or mounth.
If \code{NULL} (default) tries to guess.}

\item{agg_time}{sub_year_col might contain also aggregate values, like
yearly sums. They are exluded based on order of factor level or occurence
in data. Default is that nothing is exluded.}
}
\value{
A data.frame
}
\description{
Convert Statfin time variable(s) to numeric or date
}
\examples{
public_debt2 <- clean_times(public_debt)
str(public_debt2)
str(clean_times(output_ind, time_format = "date"))
employment_q2 <- clean_times(x = employment_q, sub_year_col = "Ajanjakso", agg_time = "Vuosikeskiarvo")
}
