% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{codes2names}
\alias{codes2names}
\title{Translates code columns to name columns}
\usage{
codes2names(.data, codes_names, to_name = names(codes_names))
}
\arguments{
\item{.data}{A data.frame or similar.}

\item{codes_names}{A named (column codes) list of named (codes) vectors (names).
For example from [px_code_name()].}

\item{to_name}{A columns to name. Default are names from [codes_names] list.}
}
\value{
A modified data.frame or similar.
}
\description{
Translates code columns to name columns
}
\examples{

x <- data.frame(a = c("a1", "a2"), b = c("b1", "b2"))
cn <- list(a = c("a1" = "first", "a2" = "second"),
           b = c("b1" = "other", "b2" = "something"))
codes2names(x, cn)
codes2names(x, cn, "a")
}
