% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_code}
\alias{extract_code}
\title{Extract a code component}
\usage{
extract_code(x, numbers_as_numeric = TRUE)
}
\arguments{
\item{x}{A character vector (or a factor).}

\item{numbers_as_numeric}{A locigal. Whether to try to convert a code to
a numeric.}
}
\value{
A character vector (or a factor).
}
\description{
This uses a regular expression to get a code from beging of a character
string. Useful to extract names from code-name variables.
}
\examples{
 extract_code("508 Mantta-Vilppula")
 extract_code("508 Mantta-Vilppula", numbers_as_numeric = FALSE)
}
\seealso{
\code{\link{extract_name}}
}
