% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/names.R
\name{make_names}
\alias{make_names}
\title{Make syntactically valid names}
\usage{
make_names(x, ...)
}
\arguments{
\item{x}{A character vector with names to change}

\item{...}{Arguments passed to \code{\link{make.names}}}
}
\value{
a character vector.
}
\description{
Try to make more convinient names than \code{\link{make.names}}.
}
\details{
Modifications:
 * replace some umlaut mark characters with a similar characters without
 * a space and punctuations replaced with "_"
 * remove duplicated, leading and trailing "_"
 * finnish with \code{\link{make.names}}
}
\examples{
make_names(c("M\\u00E4\\u00E4r\\u00E4", "Regional code"))
  make_names("Hello, world!")
}

