% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{make_names}
\alias{make_names}
\title{Make syntactically valid names}
\usage{
make_names(x, to_lower = TRUE, rename_values = FALSE, ...)
}
\arguments{
\item{x}{a character vector (of factor) with names to change}

\item{to_lower}{a locigal whether to turn names to lower case.}

\item{rename_values}{a locigal whether to rename last column (values column) as "values".}

\item{...}{Arguments passed to \code{\link{make.names}}.}
}
\value{
a character vector (or factor).
}
\description{
Try to make more convinient names than \code{\link{make.names}}.
}
\details{
Modifications:
 * replace some umlaut mark characters with a similar characters without
 * a space and punctuations replaced with "_"
 * remove duplicated, leading and trailing "_"
 * finnish with \code{\link{make.names}}
}
\examples{
  make_names(c("M\u00E4\u00E4r\u00E4", "Regional code"))
  make_names("Hello, world!")
}
