% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonal_adj}
\alias{seasonal_adj}
\alias{sa_series}
\alias{trend_series}
\title{Easy and error-resistant wrap-up for seasonal adjustment}
\usage{
seasonal_adj(x, time, series = "sa", outlier.types = "ao", ...)

sa_series(x, time, ...)

trend_series(x, time, ...)
}
\arguments{
\item{x}{a numeric vector to seasonal adjust.}

\item{time}{a date vector.}

\item{series}{a series to return. "sa" for seasonal adjusted and "trend"
for trend.}

\item{outlier.types}{an outlier parameter passed to
\code{\link[seasonal]{seas}}. Default "ao", only additive outliers
differs from \code{seas} default.}

\item{...}{parameters passed to \code{\link[seasonal]{seas}}.}
}
\description{
Seasonal adjustment using \code{\link[seasonal]{seasonal-package}}. Functions
take numeric and date vector and return numeric vector. If adjustment
end in error, NA vector returned with warning.
}
\details{
Intended specially to use in pipe with \code{\link[dplyr]{mutate}}.
}
\section{Functions}{
\itemize{
\item \code{sa_series}: seasonal adjusted series

\item \code{trend_series}: trend of seasonal adjusted series
}}

\examples{
y <- sa_series(x = AirPassengers, time = zoo::as.Date(AirPassengers))
}
