% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonal_adj}
\alias{seasonal_adj}
\alias{sa_series}
\alias{trend_series}
\title{Easy and error-resistant wrap-up for seasonal adjustment}
\usage{
seasonal_adj(x, time, ..., series = "sa")

sa_series(x, time, ...)

trend_series(x, time, ...)
}
\arguments{
\item{x}{a numeric vector to seasonal adjust.}

\item{time}{a date vector.}

\item{...}{parameters passed to \code{\link[seasonal]{seas}}.}

\item{series}{a series to return. "sa" for seasonal adjusted and "trend"
for trend.}
}
\description{
Seasonal adjustment using \code{\link[seasonal]{seasonal-package}}. Functions
take numeric and date vector and return numeric vector. If adjustment
end in error, NA vector returned with warning.
}
\details{
Intended specially to use in pipe with \code{\link[dplyr]{mutate}}.
}
\section{Functions}{
\itemize{
\item \code{sa_series}: seasonal adjusted series

\item \code{trend_series}: trend of seasonal adjusted series
}}

\examples{
org <- AirPassengers \%>\%
  tibble(time = lubridate::date_decimal(as.numeric(time(.))), values = .)

y <- sa_series(x = org$values, time = org$time)
}
