% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sf_get_class.R
\name{sf_get_class}
\alias{sf_get_class}
\title{Download statistical classifications from the Statistics Finland.}
\usage{
sf_get_class(class, year, lang = "fi")
}
\arguments{
\item{class}{a a name of the classification.}

\item{year}{a year of the classification.}

\item{lang}{a language code. Available "fi", "en", "sv".}
}
\value{
a data.frame.
}
\description{
Download a data.frame of classifications from the Statistics Finland webpage.
The data.frame includes codes and names for a classifications.
}
\details{
For all available classifications see
\url{http://tilastokeskus.fi/meta/luokitukset/index_en.html} (for english)
and \url{http://tilastokeskus.fi/meta/luokitukset/} (for finnish).

Some of available classification include:

Regional
\itemize{
  \item kunta
  \item seutukunta
  \item maakunta
}

Economic
\itemize{
  \item toimiala
  \item cpa
}
}
\examples{
x <- sf_get_class("kunta", 2014)
y <- sf_get_class("seutukunta", 2013)
z <- sf_get_class("maakunta", 2013)
t <- sf_get_class("toimiala", 2008, lang = "en")
}

