% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_get_classifications.R
\name{sf_get_class}
\alias{sf_get_class}
\title{Download statistical classifications from the Statistics Finland.}
\usage{
sf_get_class(class, year, lang = "fi", as_factors = default.stringsAsFactors())
}
\arguments{
\item{class}{a a name of the classification.}

\item{year}{a year of the classification.}

\item{lang}{a language code. Available "fi", "en", "sv".}

\item{as_factors}{a logical. Should columns be converted to factors. The
default is from \code{\link{options}}("stringsAsFactors").}
}
\value{
a data.frame.
}
\description{
Download a data.frame of classifications from the Statistics Finland webpage.
The data.frame includes codes and names for a classifications.
}
\details{
For all available classifications see
\url{http://tilastokeskus.fi/meta/luokitukset/index_en.html} (for english)
and \url{http://tilastokeskus.fi/meta/luokitukset/} (for finnish).

Some of available classification include:

Regional
\itemize{
  \item kunta
  \item seutukunta
  \item maakunta
  \item kuntaryhmitys
  \item tyossakayntial
}

Economic
\itemize{
  \item toimiala
  \item cpa
}
}
\examples{
x <- sf_get_class(class = "kunta", year = 2016)
y <- sf_get_class("seutukunta", 2013)
z <- sf_get_class("maakunta", 2016)
t <- sf_get_class("toimiala", 2008, lang = "en")
}
\seealso{
Other sf_class: 
\code{\link{sf_get_class_key}()},
\code{\link{sf_get_reg_key}()}
}
\concept{sf_class}
