% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_recode.R
\name{sf_recode}
\alias{sf_recode}
\alias{sf_code2name}
\alias{sf_name2code}
\title{Turns codes to names or vice versa in Statistics Finland classifications}
\usage{
sf_recode(x, from, to, ...)

sf_code2name(x, ...)

sf_name2code(x, ...)
}
\arguments{
\item{x}{A vector for recoding.}

\item{from}{A column name or position for orginal classification
in the classification data.frame. See \code{\link{sf_get_class}}.}

\item{to}{A column name or position for desidered classification
in the classification data.frame. See \code{\link{sf_get_class}}.}

\item{...}{futher parameters to \code{\link{sf_get_class}} to select classification.}
}
\description{
Recode statfin classifications from numeric codes to names or vice versa.
\code{sf_code2name} and \code{sf_name2code} and shortcuts to main
\code{sf_recode} function.
}
\details{
It is best to use column postions for \code{from} and \code{to}.
The first should be a numeric code and the second a name.
}
\section{Functions}{
\itemize{
\item \code{sf_code2name()}: From code to name.

\item \code{sf_name2code()}: From name to code.

}}
