% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_urls.R
\name{statfi_url}
\alias{statfi_url}
\title{Statfi url}
\usage{
statfi_url(
  ...,
  with_base_url = TRUE,
  lang = "fi",
  .base_url = "https://statfin.stat.fi/PXWeb/api/v1"
)
}
\arguments{
\item{...}{Character vectors.}

\item{with_base_url}{logical, whether to add the base of url to the output.
Defaults to \code{TRUE}}

\item{.base_url}{A base url for statfi.}
}
\description{
Gives full statfi url
}
\examples{
  statfi_url("StatFin", "muutl/statfin_muutl_pxt_119z.px")

}
